refactor: 페이지네이션 관련 로직 훅으로 분리
